The dimensions of feature/label matrices are now of type uint32.

+++ python/boomer/common/cpp/input_data.cpp
@@ -1,7 +1,7 @@
 #include "input_data.h"
 
 
-AbstractLabelMatrix::AbstractLabelMatrix(intp numExamples, intp numLabels) {
+AbstractLabelMatrix::AbstractLabelMatrix(uint32 numExamples, uint32 numLabels) {
     numExamples_ = numExamples;
     numLabels_ = numLabels;
 }
@@ -10,7 +10,7 @@ AbstractLabelMatrix::~AbstractLabelMatrix() {
 
 }
 
-AbstractRandomAccessLabelMatrix::AbstractRandomAccessLabelMatrix(intp numExamples, intp numLabels)
+AbstractRandomAccessLabelMatrix::AbstractRandomAccessLabelMatrix(uint32 numExamples, uint32 numLabels)
     : AbstractLabelMatrix(numExamples, numLabels) {
 
 }
@@ -19,7 +19,7 @@ uint8 AbstractRandomAccessLabelMatrix::getLabel(intp exampleIndex, intp labelInd
     return 0;
 }
 
-DenseLabelMatrixImpl::DenseLabelMatrixImpl(intp numExamples, intp numLabels, const uint8* y)
+DenseLabelMatrixImpl::DenseLabelMatrixImpl(uint32 numExamples, uint32 numLabels, const uint8* y)
     : AbstractRandomAccessLabelMatrix(numExamples, numLabels) {
     y_ = y;
 }
@@ -33,7 +33,8 @@ uint8 DenseLabelMatrixImpl::getLabel(intp exampleIndex, intp labelIndex) {
     return y_[i];
 }
 
-DokLabelMatrixImpl::DokLabelMatrixImpl(intp numExamples, intp numLabels, std::shared_ptr<BinaryDokMatrix> dokMatrixPtr)
+DokLabelMatrixImpl::DokLabelMatrixImpl(uint32 numExamples, uint32 numLabels,
+                                       std::shared_ptr<BinaryDokMatrix> dokMatrixPtr)
     : AbstractRandomAccessLabelMatrix(numExamples, numLabels) {
     dokMatrixPtr_ = dokMatrixPtr;
 }
